Fix typo - remove "bye" in sample output.
When the user adds the word "bye" to their program upon not finding the characters "av", the exercise is tested to be incorrect. When the user removes this functionality however, their solution is said to be correct. Because of this, the word "bye" should not be in the output as it misleads the user.