8-2 two issues
Exercise : "Abbreviations"
No abbreviation is added with an explanation for "lol". While running the for loop (ln.943), the conditional in ln.944 should return false, so the part with "lol" should not reach the .findExplanationFor("lol") which would return 'null'. That is why the sample output on ln.958 should contain "lol" instead of "null".

Exercise : "I Owe You"
In sample output we should have the corrected text as on line 1425 we call .setSum("Michael", 30) so on line 1428, Michael should be found and his sum (30.0) should be printed instead of 0.0.